% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatcat-package.R
\docType{package}
\name{flatcat-package}
\alias{flatcat}
\alias{flatcat-package}
\title{flatcat: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/flatcat}
  \item Report bugs at \url{https://github.com/DavisVaughan/flatcat/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

}
\keyword{internal}
